Rework hierarchy to remove duplicated code
