+++ src/main/java/me/s3ns3iw00/jcommands/argument/SubArgument.java
@@ -1,5 +1,7 @@
 package me.s3ns3iw00.jcommands.argument;
 
+import org.javacord.api.interaction.SlashCommandOptionType;
+
 import java.util.LinkedList;
 
 /**
@@ -8,6 +10,20 @@ import java.util.LinkedList;
  */
 public abstract class SubArgument extends Argument {
 
-    public abstract LinkedList<Argument> getArguments();
+    private final LinkedList<Argument> arguments = new LinkedList<>();
 
+    /**
+     * Constructs the argument with the default requirements
+     *
+     * @param name        the argument's name
+     * @param description the argument's description
+     * @param type        the type of the input value
+     */
+    public SubArgument(String name, String description, SlashCommandOptionType type) {
+        super(name, description, type);
+    }
+
+    public LinkedList<Argument> getArguments() {
+        return arguments;
+    }
 }
